Extract comment date formatting
Moves formatDate and formatDateTime out of Comment.js so that other
review components can render timestamps the same way.
